+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Bug fixes
 - PIM-3406: Fix boolean filter on Mongo implementation
-- PIM-3430: Fix doctrine issue on prices when skip item during product import
+- PIM-3430: Fix doctrine issue on prices when skip an item during the product import
+- PIM-3358: Fix sprintf issue in an exception which prevents doctrine writer to deal with anything else than an object
 
 # 1.2.12 (2014-11-13)